+++ src/main/java/org/co2/kanban/rest/user/ApplicationUserController.java
@@ -5,22 +5,12 @@
  */
 package org.co2.kanban.rest.user;
 
-import java.util.List;
-import org.co2.kanban.repository.project.Project;
-import org.co2.kanban.repository.project.ProjectRepository;
-import org.co2.kanban.repository.task.Task;
-import org.co2.kanban.repository.task.TaskRepository;
+import java.security.Principal;
 import org.co2.kanban.repository.user.ApplicationUser;
 import org.co2.kanban.repository.user.ApplicationUserRepository;
-import org.co2.kanban.repository.user.ApplicationUserRole;
-import org.co2.kanban.rest.project.ProjectAssembler;
-import org.co2.kanban.rest.project.ProjectResource;
-import org.co2.kanban.rest.project.member.MemberAssembler;
-import org.co2.kanban.rest.project.member.MemberResource;
-import org.co2.kanban.rest.project.task.TaskAssembler;
-import org.co2.kanban.rest.project.task.TaskResource;
+import org.co2.kanban.rest.error.BusinessException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.hateoas.PagedResources;
@@ -30,11 +20,14 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
@@ -49,44 +42,33 @@ public class ApplicationUserController {
     private ApplicationUserRepository repository;
 
     @Autowired
-    private MemberAssembler memberAssembler;
-
-    @Autowired
-    private TaskRepository taskRepositoy;
-
-    @Autowired
-    private ProjectRepository projectRepositoy;
-
-    @Autowired
     private UserAssembler userAssembler;
 
     @Autowired
     private PagedResourcesAssembler<ApplicationUser> pagedAssembler;
 
-    @Autowired
-    private ProjectAssembler projectAssembler;
+    private final BCryptPasswordEncoder bcryptEncoder = new BCryptPasswordEncoder();
 
-    @Autowired
-    private TaskAssembler taskAssembler;
-
-    @Autowired
-    private PagedResourcesAssembler<Project> pagedProjectAssembler;
-
-    @Autowired
-    private PagedResourcesAssembler<Task> pagedTaskAssembler;
-
-    @RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public PagedResources<UserResource> page(Pageable p) {
-        return pagedAssembler.toResource(repository.findAll(p), userAssembler);
+    @RequestMapping(params = {"page", "size"}, method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+    public PagedResources<UserResource> page(
+            @RequestParam(name = "page") Integer page,
+            @RequestParam(name = "size", required = false) Integer size) {
+        Pageable pageable = new PageRequest(page, size);
+        return pagedAssembler.toResource(repository.findAll(pageable), userAssembler);
     }
 
-    @RequestMapping(value = "/find/{username}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public List<UserResource> findByUsername(@PathVariable("username") String username) {
-        return userAssembler.toResources(repository.findByUsernameContaining(username));
+    @RequestMapping(params = {"search"}, method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+    public Iterable<UserResource> search(@RequestParam(name = "search") String search) {
+        return userAssembler.toResources(repository.findByUsernameContains(search));
     }
 
     @RequestMapping(method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<UserResource> create(@RequestBody ApplicationUser newUser) {
+        if (repository.checkExistUsername(newUser.getUsername())) {
+            throw new BusinessException(HttpStatus.CONFLICT, "user.error.conflict.username");
+        }
+        String passwordDigest = bcryptEncoder.encode(newUser.getPassword());
+        newUser.setPassword(passwordDigest);
         ApplicationUser user = repository.save(newUser);
         MultiValueMap<String, String> headers = new HttpHeaders();
         headers.add(HttpHeaders.LOCATION, linkTo(methodOn(ApplicationUserController.class).get(user.getId())).toString());
@@ -100,33 +82,21 @@ public class ApplicationUserController {
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity delete(@PathVariable("id") Long userId) {
+    public ResponseEntity delete(@AuthenticationPrincipal Principal user, @PathVariable("id") Long userId) {
+        ApplicationUser currentUser = repository.findByUsername(user.getName());
+        if (currentUser.getId().equals(userId)) {
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+        }
         repository.delete(userId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @RequestMapping(value = "/{id}/member", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<MemberResource>> memberOf(@PathVariable("id") Long userId) {
-        return new ResponseEntity<>(memberAssembler.toResources(repository.findOne(userId).getMembers()), HttpStatus.OK);
+    @RequestMapping(value = "/{id}", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity update(@PathVariable("id") Long userId, @RequestBody ApplicationUser user) {
+        ApplicationUser updatedUser = repository.findOne(userId);
+        updatedUser.setApplicationRole(user.getApplicationRole());
+        updatedUser.setEmail(user.getEmail());
+        repository.save(updatedUser);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
-
-    @RequestMapping(value = "/{id}/task", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public PagedResources<TaskResource> listTask(@PathVariable("id") Long userId, Pageable page) {
-        ApplicationUser appUser = repository.findOne(userId);
-        Page<Task> tasks = taskRepositoy.findByStateCloseStateFalseAndAssigneeUserOrBackupUser(appUser, appUser, page);
-        return pagedTaskAssembler.toResource(tasks, taskAssembler);
-    }
-
-    @RequestMapping(value = "/{id}/project", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public PagedResources<ProjectResource> listProject(@PathVariable("id") Long userId, Pageable page) {
-        ApplicationUser appUser = repository.findOne(userId);
-        Page<Project> projects;
-        if (appUser.getApplicationRole().equals(ApplicationUserRole.ADMIN)) {
-            projects = projectRepositoy.findAll(page);
-        } else {
-            projects = projectRepositoy.findByMembersUser(appUser, page);
-        }
-        return pagedProjectAssembler.toResource(projects, projectAssembler);
-    }
-
 }
